@@ -1,6 +1,7 @@
 {
   "name": "demo",
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
@@ -12,11 +13,11 @@
     "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "@types/react": "18.0.15",
-    "@types/react-dom": "18.0.6",
-    "@vitejs/plugin-react": "4.1.1",
-    "typescript": "5.2.2",
-    "vite": "4.5.0"
+    "@types/react": "18.2.43",
+    "@types/react-dom": "18.2.17",
+    "@vitejs/plugin-react": "4.2.1",
+    "typescript": "5.3.3",
+    "vite": "5.0.6"
   },
   "license": "MIT"
 }
